crates/opentelemetry/examples/pyo3-opentelemetry-lib/pyproject.toml
[project]
name = "pyo3_opentelemetry_lib"
version = "0.0.1"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
description = "Propagation of OpenTelemetry context from Python to Rust"
readme = "README.md"
authors = [
    { name = "Rigetti Computing", email = "softapps@rigetti.com" },
    { name = "Eric Hulburd", email = "ehulburd@rigetti.com" },
]

dependencies = [
    "opentelemetry-api ~= 1.39",
    "opentelemetry-sdk ~= 1.39",
    "opentelemetry-proto ~= 1.39",
    "typing-extensions ~= 4.15",
]

[tool.poetry.group.test]
optional = true
[tool.poetry.group.test.dependencies]
grpcio = "^1.59.0"
pytest = "^9.0.0"
pytest-asyncio = "^1.3.0"
pytest-forked = "^1.6.0"
pytest-timeout = "^2.4.0"

[tool.poetry.group.lint]
optional = true
[tool.poetry.group.lint.dependencies]
black = "^25.12.0"
pyright = "^1.1.408"
ruff = "^0.14.11"
mypy = "^1.19.1"

[tool.poetry.group.dev]
optional = true
include-groups = [ "test", "lint" ]
[tool.poetry.group.dev.dependencies]
maturin = "^1.9.4"

[tool.poetry]
package-mode = false

[build-system]
requires = ["maturin>=1.9.4,<2.0.0"]
build-backend = "maturin"

[tool.maturin]
bindings = "pyo3"
compatibility = "linux"
sdist-include = ["README.md"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
preview= true
extend-exclude = '''
(
    "pyo3_opentelemetry_lib\/__init__.py",
)
'''

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "grpc",
    "grpc.aio",
]
ignore_missing_imports = true

[tool.pyright]
exclude = [
    "pyo3_opentelemetry_lib/__init__.py",
    ".venv"
]
reportUnsupportedDunderAll = false

[tool.ruff]
src = ["pyo3_opentelemetry_lib", "tests"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 120

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001"
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.pytest.ini_options]
timeout = 900

